requirements.txt
slack-bolt==1.21.3
gspread==6.2.1
google-auth==2.48.0
anthropic==0.52.0
